refactor formatting to viz utils
