+++ src/vivarium_testing_utils/automated_validation/visualization/plot_utils.py
